Explore REST API for the service
In this commit, I'm exploring REST API for this service,
When I'm ok with /resources, /viewers, .../entitlements
I'm not sure what to think about .../available-resources

it's still "resources" but in the context of the viewer.
I was thinking about /viewers/{id}/resources but it doesn't seem self-descriptive as it should be.